+++ Team8_Project/Source/Team8_Project/Inventory/ConsumableItemRow.h
@@ -2,8 +2,8 @@
 
 #include "CoreMinimal.h"
 #include "Engine/DataTable.h"
-#include "InventoryType.h" 
-#include "BaseItem.h"
+#include "../Inventory/InventoryType.h" 
+#include "../BaseItem.h"
 #include "ConsumableItemRow.generated.h"
 
 UENUM(BlueprintType)
