+++ Team8_Project/Source/Team8_Project/Inventory/WeaponBase.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "BaseItem.h"
+#include "../BaseItem.h"
 #include "WeaponBase.generated.h"
