+++ Team8_Project/Source/Team8_Project/MyCharacter.cpp
@@ -1,8 +1,8 @@
 ﻿#include "MyCharacter.h"
 #include "MyPlayerController.h"
-#include "InventoryComponent.h"
-#include "InventorySubsystem.h"
-#include "InventoryWidget.h"
+#include "Inventory/InventoryComponent.h"
+#include "Inventory/InventorySubsystem.h"
+#include "Inventory/InventoryWidget.h"
 #include "Components/CapsuleComponent.h"
 #include "BaseItem.h"
 #include "GameFramework/SpringArmComponent.h"
